requirements.txt
requests
lxml
flask
flask-sqlalchemy
pymysql
PyYAML
redis